+++ crates/ukis_h3cellstorepy/tests/clickhouse/__init__.py
@@ -13,7 +13,7 @@ def elephant_schema(tableset_name="okavango_delta", temporal_partitioning="month
     csb.h3_partitioning(h3_partitioning, **kw)
     csb.add_column("is_valid", "UInt8", compression_method=CompressionMethod("gorilla"))
     csb.add_aggregated_column("elephant_density", "Float32", "RelativeToCellArea")
-    csb.add_aggregated_column("some_category", "UInt8", "SetNull", nullable=True)
+    csb.add_aggregated_column("some_category", "UInt8", "SetNullOnConflict", nullable=True)
     schema = csb.build()  # raises when the schema is invalid / missing something
     assert schema is not None
     #print(schema.to_json_string())
@@ -42,11 +42,13 @@ def setup_elephant_schema_with_data(clickhouse_grpc_endpoint, clickhouse_testdb_
     center_point = (20.0, 10.0)
     # uncompacted disk
     disk = h3.k_ring(h3.geo_to_h3(center_point[1], center_point[0], schema.max_h3_resolution), 10).astype(np.uint64)
+    cat1 = np.ones(int(len(disk) / 2)) * 23
+    cat2 = np.ones(len(disk)- len(cat1)) * 12
     df = pl.DataFrame({
         "h3index": disk,
         "is_valid": np.ones(len(disk)),
         "elephant_density": np.ones(len(disk)) * 4,
-        "some_category": np.ones(len(disk)) * 23
+        "some_category": np.concatenate((cat1, cat2))
     })
 
     # write to db - this performs auto-compaction